Added authorization check and improve services functionality.

+++ src/main/java/com/post/jibberjabberposts/service/impl/PostServiceImpl.java
@@ -5,8 +5,11 @@ import com.post.jibberjabberposts.dto.PostDto;
 import com.post.jibberjabberposts.dto.ReplyCreationDto;
 import com.post.jibberjabberposts.model.Post;
 import com.post.jibberjabberposts.model.Reply;
+import com.post.jibberjabberposts.model.User;
 import com.post.jibberjabberposts.repository.PostRepository;
+import com.post.jibberjabberposts.repository.UserRepository;
 import com.post.jibberjabberposts.service.PostService;
+import com.post.jibberjabberposts.service.UserService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -14,6 +17,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 import java.util.logging.Logger;
 
@@ -22,14 +26,24 @@ public class PostServiceImpl implements PostService {
 
     private final PostRepository postRepository;
 
+    private final UserService userService;
+
     private final Logger logger = Logger.getLogger(PostServiceImpl.class.getName());
 
-    public PostServiceImpl(PostRepository postRepository) {
+    public PostServiceImpl(PostRepository postRepository, UserService userService) {
         this.postRepository = postRepository;
+        this.userService = userService;
     }
 
     @Override
     public PostDto createPost(PostCreationDto postCreationDto) {
+        User user = userService.getCurrentUser();
+        if(user == null) {
+            throw new IllegalArgumentException("User not found");
+        }
+        if(user.getId() != postCreationDto.getAuthorId()) {
+            throw new IllegalArgumentException("User cannot create post for other user");
+        }
         Post post = Post.builder()
                 .content(postCreationDto.getContent())
                 .authorId(postCreationDto.getAuthorId())
@@ -49,17 +63,29 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public Page<PostDto> getPostsByUser(UUID userId, int page, int size) {
-        Pageable pageable = PageRequest.of(page, size);
-        Page<Post> posts = postRepository.findAllByAuthorId(userId, pageable);
-        return posts.map(PostDto::from);
+        User user = userService.getUserById(userId);
+        if (user == null) {
+            throw new IllegalArgumentException("User not found");
+        } else {
+            Pageable pageable = PageRequest.of(page, size);
+            Page<Post> posts = postRepository.findAllByAuthorId(userId, pageable);
+            return posts.map(PostDto::from);
+        }
     }
 
     @Override
     public PostDto createReply(UUID postId, ReplyCreationDto replyCreationDto) {
+        User user = userService.getCurrentUser();
+        if(user == null) {
+            throw new IllegalArgumentException("User not found");
+        }
+        if(user.getId() != replyCreationDto.getAuthorId()) {
+            throw new IllegalArgumentException("User cannot create reply for other user");
+        }
         Post post = postRepository.findById(postId).orElseThrow(() -> new RuntimeException("Post not found"));
         Reply reply = Reply.builder()
                 .content(replyCreationDto.getContent())
-                .authorId(replyCreationDto.getAuthorId())
+                .authorId(user.getId())
                 .build();
         post.getReplies().add(reply);
         logger.info("New Reply");
@@ -72,6 +98,13 @@ public class PostServiceImpl implements PostService {
     public void deletePost(UUID postId) {
         logger.info("Deleting Post: " + postId);
         Post post = postRepository.findById(postId).orElseThrow(() -> new RuntimeException("Post not found"));
+        User user = userService.getCurrentUser();
+        if(user == null) {
+            throw new IllegalArgumentException("User not found");
+        }
+        if(user.getId() != post.getAuthorId()) {
+            throw new IllegalArgumentException("User cannot delete post for other user");
+        }
         postRepository.delete(post);
         logger.info("Post deleted");
     }
